+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - Restrict utility menu output to the parent level, fixes #61.
 - Prevent child menu items from being added to the Utility menu, fixes #61.
-- Delete child menu items from the Utility menu on save operation, fixes #61.
+- Reset nested Utility menu items to the top/parent level on save, fixes #61.
+- Add admin notice stating the utility nav restrictions, fixes #61.
 
 ## 2.5.2